pref: 添加灵动岛性能模式优化maxexpand界面性能
